Regular (partly working) phpvcr fixtures

+++ Tests/Functional/AbstractBaseFunctionTest.php
@@ -71,6 +71,9 @@ abstract class AbstractBaseFunctionTest extends WebTestCase
             $di = new \RecursiveDirectoryIterator($path, \FilesystemIterator::SKIP_DOTS);
             $ri = new \RecursiveIteratorIterator($di, \RecursiveIteratorIterator::CHILD_FIRST);
             foreach ($ri as $file) {
+                if ($file->getFilename() === '.gitignore') {
+                    continue;
+                }
                 $file->isDir() ? rmdir($file) : unlink($file);
             }
         } else {
